% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMDist.R
\name{CMDist}
\alias{CMDist}
\title{Concept Mover's Distance Function}
\usage{
CMDist(
  dtm,
  cw = NULL,
  cd = NULL,
  wv,
  method = "cosine",
  scale = TRUE,
  parallel = FALSE,
  threads = 2
)
}
\arguments{
\item{Function}{requires a document-term matrix and a matrix of word embeddings.}
}
\description{
The function outputs a dataframe with Concept Mover's Distances for each document.
}
\examples{
cm.dists <- CMDist(dtm, cw = "death", wv = wordvectors, scale = TRUE)
}
\references{
\url{https://journals.sagepub.com/doi/10.1177/2378023119827674}
}
