% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kusnerRWMD.R
\name{KusnerRelaxedWordMoversDistance}
\alias{KusnerRelaxedWordMoversDistance}
\alias{kusnerRWMD}
\title{Creates model which can be used for calculation of "relaxed word movers distance".}
\description{
Creates model which can be used for calculation of "relaxed word movers distance".

Creates model which can be used for calculation of "relaxed word movers distance".
}
\section{Super class}{
\code{\link[CMDist:distance_model]{CMDist::distance_model}} -> \code{RWMD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KusnerRelaxedWordMoversDistance$new()}}
\item \href{#method-dist2}{\code{KusnerRelaxedWordMoversDistance$dist2()}}
\item \href{#method-clone}{\code{KusnerRelaxedWordMoversDistance$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KusnerRelaxedWordMoversDistance$new(
  wv,
  method = c("cosine", "euclidean"),
  normalize = TRUE,
  progressbar = interactive()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dist2"></a>}}
\subsection{Method \code{dist2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KusnerRelaxedWordMoversDistance$dist2(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KusnerRelaxedWordMoversDistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[CMDist:distance_model]{CMDist::distance_model}} -> \code{RWMD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{kusnerRWMD$new()}}
\item \href{#method-dist2}{\code{kusnerRWMD$dist2()}}
\item \href{#method-clone}{\code{kusnerRWMD$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kusnerRWMD$new(
  wv,
  method = c("cosine", "euclidean"),
  normalize = TRUE,
  progressbar = interactive()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dist2"></a>}}
\subsection{Method \code{dist2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kusnerRWMD$dist2(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kusnerRWMD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
